ctf: generate CTF information for the kernel
This introduces a new config option, CONFIG_CTF, which causes the
compiler to be run with -gctf to emit information on the kernel's types
and global variables, and a tool, ctfarchive, which runs whenever
'make ctf' is run, deduplicating that information into a mmappable
type archive named vmlinux.ctfa, which is installed at 'make ctf_install'
time into /lib/modules/$(uname -r)/kernel/.  You can view the contents of
this archive via

objdump --ctf --ctf-parent=shared_ctf /lib/modules/$(uname -r)/kernel/vmlinux.ctfa

from binutils 2.41+ (but only binutils 2.35+ is needed for generation).

Out-of-tree modules cannot participate in this mechanism since vmlinux.ctfa
is already written: so, as happens for normal userspace binaries, the compiler
emits .ctf sections into the modules themselves, and the linker deduplicates
them.  strip does not strip such sections out, but additional care should be
taken not to strip such sections into debug RPMs (they are small enough that
this should not be a problem).

For in-tree modules and the core kernel, ctfarchive uses the same code (in
libctf) to do the deduplication into vmlinux.ctfa as the linker does for
out-of-tree modules and userspace code compiled with -gctf: there is
no code duplication.

Within the ctfa file, the type information is divided into a shared
repository, containing all types used by more than one module, CTF for
the core kernel, and separate CTF for each module built, whether or not
this module has been compiled in or not: if a file *could* be built as a
module, it will be considered to be a module from the perspective of CTF
file emission.  This ensures that external consumers such as DTrace
always find types for a given module in the same place, regardless of
the local kernel configuration, as long as that module is present at
all, assisting in portability of D scripts between installations.

This process relies on modules.builtin.objs (added by the commits
immediately preceding this one), which maps from the thin archives that make
up built-in modules to their constituent object files.

We use separate targets for CTF deduplication and installation so that
you don't pay the cost of cross-module CTF deduplication while iterating
on kernel development, and so that ordinary kernel make install targets
don't surprise scripts with extra files.  I could be convinced that it
makes more sense to install the CTF via make install iff it exists.

One downside of separating make modules_install and make ctf_install
is that since make modules_install deletes /lib/modules/$(uname -r),
make ctf_install only has an effect if you run it after make modules_install,
not before.

This introduces new kernel build-time dependencies on a sufficiently new
binutils, a CTF-capable GCC, and libctf: if any of these is not found,
CONFIG_CTF is disabled and all this machinery has no effect.  Even if it
is turned on, in the absence of 'make ctf', the CTF is deduplicated into
the modules and then stripped from them and discarded; and in the absence
of 'make ctf_install', the deduplicated vmlinux.ctfa is not installed.

Source-wise, outside of various places turning CTF off where not desired
(mostly the vDSO), and deleting the CTF from the final modules and
kernel, there is hardly any impact to the existing source tree: the new
rules are mostly centralized in a new scripts/Makefile.ctfa and a small
utility 'ctfarchive' to do the final deduplication.

Orabug: 29891866
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>